Move runtime out of main into a module
